+++ index.html
@@ -60,7 +60,7 @@
 
                   <!-- Изменение глубины эффекта, накладываемого на изображение -->
                   <fieldset class="img-upload__effect-level  effect-level">
-                    <input class="effect-level__value" type="number" step="any" name="effect-level" value="">
+                    <input class="effect-level__value" type="number" step="any" name="effect-level" id="effect-level" value="">
                     <div class="effect-level__slider"></div>
                   </fieldset>
 
@@ -249,6 +249,7 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
+    <script src="vendor/nouislider/nouislider.js"></script>
     <script src="vendor/pristine/pristine.min.js"></script>
     <script src="js/main.js" type="module"></script>
