+++ js/edit-form-open-close.js
@@ -1,17 +1,12 @@
 import { isEscapeKey, showModal, closeModal } from './utils.js';
-import {
-  scaleSmallerButtonEl,
-  scaleBiggerButtonEl,
-  resetScale,
-  onScaleSmallerClick,
-  onScaleBiggererClick
-} from './edit-scale-inp.js';
+import { uploadImgFormEl, onEditFormSaubmit } from './validate-form.js';
+import { editImgEl, resetScale, onScaleSmallerClick, onScaleBiggererClick, resetEffect, onEffectChange } from './edit-upload-img.js';
 
-const uploadImgFormEl = document.querySelector('#upload-select-image');
 const uploadImgButtonEl = uploadImgFormEl.querySelector('#upload-file');
-const editImgEl = uploadImgFormEl.querySelector('.img-upload__overlay');
 const closeEditImgButtonEl = editImgEl.querySelector('.img-upload__cancel');
-const editorPreview = editImgEl.querySelector('.img-upload__preview').querySelector('img');
+const scaleSmallerButtonEl = editImgEl.querySelector('.scale__control--smaller');
+const scaleBiggerButtonEl = editImgEl.querySelector('.scale__control--bigger');
+const effectsListEl = editImgEl.querySelector('.effect-level__slider');
 
 const showFullscreenEditor = () => {
   // console.log(uploadImgButtonEl.value); - как-то подставим это фото в editorPreview
@@ -19,16 +14,21 @@ const showFullscreenEditor = () => {
   closeEditImgButtonEl.addEventListener('click', onCloseButtonClick);
   scaleSmallerButtonEl.addEventListener('click', onScaleSmallerClick);
   scaleBiggerButtonEl.addEventListener('click', onScaleBiggererClick);
+  effectsListEl.addEventListener('click', onEffectChange);
+  uploadImgFormEl.addEventListener('submit', onEditFormSaubmit);
 };
 
 const closeFullscreenEditor = () => {
   uploadImgFormEl.reset();
   closeModal(editImgEl);
   resetScale();
+  resetEffect();
   document.removeEventListener('keydown', onEscKeydown);
   closeEditImgButtonEl.removeEventListener('click', onCloseButtonClick);
   scaleSmallerButtonEl.removeEventListener('click', onScaleSmallerClick);
   scaleBiggerButtonEl.removeEventListener('click', onScaleBiggererClick);
+  effectsListEl.removeEventListener('click', onEffectChange);
+  uploadImgFormEl.removeEventListener('submit', onEditFormSaubmit);
 };
 
 function onEscKeydown(evt) {
@@ -53,8 +53,3 @@ function onuploudButtionClick(evt) {
 }
 
 uploadImgButtonEl.addEventListener('change', onuploudButtionClick);
-
-export {
-  uploadImgFormEl,
-  editorPreview
-};
